+++ src/composables/useSessions.ts
@@ -1,27 +1,7 @@
 import { isEqual } from 'lodash'
 import { computed, reactive, ref, type Ref, toRaw } from 'vue'
 import { clone } from '@utils/app'
-import { Storage } from '@utils/storage'
-
-export type Layout = 'hug' | 'fill' | 'nav' // | 'custom'
-export type Setting = 'layout' | string
-
-export interface Session {
-  id: string
-  title: string
-  urls: string[]
-  settings: Record<Setting, any>
-}
-
-export const Sessions = {
-  async get (): Promise<Session[]> {
-    return Storage.get('sessions', [])
-  },
-
-  async set (value: Session[]) {
-    void Storage.set('sessions', value)
-  },
-}
+import { type Session, Sessions } from '@utils/session'
 
 /**
  * Store for the saved-session list.
